raspberry pi fix

+++ Huebner Inverter.sh
@@ -23,21 +23,27 @@ checkUSB()
     done
 }
 
-if [[ $(type -p php) ]]; then
-    
-    for file in ./Linux/*; do
-        chmod +x "$file"
-    done
+if [ -f /usr/bin/minicom ]; then
 
-    echo "Running as sudo ..."
-    
-    sudo killall php
-    sudo $(type -p php) -S 0.0.0.0:8080 -t "$(dirname "$0")/Web/" &
-    sleep 4
-    
-    checkUSB
+    if [[ $(type -p php) ]]; then
+        
+        for file in ./Linux/*; do
+            chmod +x "$file"
+        done
+
+        echo "Running as sudo ..."
+        
+        sudo killall php
+        sudo php -S 0.0.0.0:8080 -t "$(dirname "$0")/Web/" &
+        sleep 4
+        
+        checkUSB
+    else
+
+        echo "PHP not Installed ...Install? [Y/n]"; read
+        ./Linux/php.sh
+    fi
 else
-
-    echo "PHP not Installed ...Install? [Y/n]"; read
-    ./Linux/php.sh
+    echo "Minicom not Installed ...Install? [Y/n]"; read
+        sudo apt-get -y install minicom
 fi
